Revert "added api url to .env and added error checking to api call:"
This reverts commit 90ec03b3efcab58485b5e26c7de8c36eb105d9d0.
