+++ src/Components/util/getData.tsx
@@ -1,23 +1,17 @@
 import checkErrorStatus from "./checkErrorStatus";
 
-const getData = async (url: any) => {
-  try {
-    const dataResponse = await fetch(url);
-    let error = checkErrorStatus(dataResponse);
-    if (error !== null) {
-      console.log("HERE");
-      throw new Error(error);
-    } else {
-      const data = await dataResponse.json();
-      return data;
-    }
-  } catch (err) {
-    if (err instanceof Error) {
-      return err.message;
-    } else {
-      return String(err);
-    }
+const getData = async (url: string) => {
+  const dataResponse = await fetch(url);
+
+  let error = checkErrorStatus(dataResponse);
+
+  if (error !== null) {
+    throw new Error(error);
   }
+
+  const data = await dataResponse.json();
+
+  console.log("data: ", data);
 };
 
 export default getData;
